+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## v1.22.0 - Unreleased
+## v1.22.0 - 2020-08-06
 
 ### Improvements
 - Use multicolor channel icons for better appearance in the dark mode